chore: separates type and const exports
