+++ src/types/constants.ts
@@ -1,0 +1,15 @@
+export const validOperators = [
+  'equals',
+  'contains',
+  'not_equals',
+  'in',
+  'all',
+  'not_in',
+  'exists',
+  'greater_than',
+  'greater_than_equal',
+  'less_than',
+  'less_than_equal',
+  'like',
+  'near',
+] as const;
